moving functions out of av_spex_the_file
