+++ src/AV_Spex/av_spex_the_file.py
@@ -20,22 +20,22 @@ from PyQt6.QtWidgets import (
     QApplication
 )
 
+from .utils import yaml_profiles
+from .utils import dir_setup
+from .utils import tools_processing
 from .utils.log_setup import logger
 from .utils.deps_setup import required_commands, check_external_dependency, check_py_version
 from .utils.find_config import config_path, command_config, yaml
-from .utils import yaml_profiles
-from .utils import dir_setup
-from .utils.generate_report import write_html_report
+from .utils.generate_report import generate_final_report
 from .utils.gui_setup import ConfigWindow, MainWindow
 from .checks import mediaconch_check
 from .checks.fixity_check import check_fixity, output_fixity
-from .checks.filename_check import is_valid_filename
 from .checks.mediainfo_check import parse_mediainfo
 from .checks.mediatrace_check import parse_mediatrace
 from .checks.exiftool_check import parse_exiftool
 from .checks.ffprobe_check import parse_ffprobe
-from .checks.embed_fixity import extract_tags, extract_hashes, embed_fixity, validate_embedded_md5
-from .checks.make_access import make_access_file
+from .checks.embed_fixity import validate_embedded_md5, process_embedded_fixity
+from .checks.make_access import make_access_file, process_access_file
 from .checks.qct_parse import run_qctparse
 
 
@@ -77,47 +77,65 @@ FILENAME_MAPPING = {
 }
 
 
-def run_command(command, input_path, output_type, output_path):
-    '''
-    Run a shell command with 4 variables: command name, path to the input file, output type (often '>'), path to the output file
-    '''
+def check_tool_metadata(tool_name, output_path, command_config):
+    """
+    Check metadata for a specific tool if configured.
+    
+    Args:
+        tool_name (str): Name of the tool
+        output_path (str): Path to the tool's output file
+        command_config (object): Configuration object with tool settings
+        
+    Returns:
+        dict or None: Differences found by parsing the tool's output, or None
+    """
+    # Mapping of tool names to their parsing functions
+    parse_functions = {
+        'exiftool': parse_exiftool,
+        'mediainfo': parse_mediainfo,
+        'mediatrace': parse_mediatrace,
+        'ffprobe': parse_ffprobe
+    }
 
-    # Get the current PATH environment variable
-    env = os.environ.copy()
-    env['PATH'] = '/usr/local/bin:' + env.get('PATH', '')
-
-    full_command = f"{command} \"{input_path}\" {output_type} {output_path}"
-
-    logger.debug(f'Running command: {full_command}\n')
-    subprocess.run(full_command, shell=True, env=env)
-
-
-def run_mediatrace_command(command, input_path):
-    '''
-    Run a shell command with 2 variables: command, path to input file
-    '''
-
-    # Get the current PATH environment variable
-    env = os.environ.copy()
-    env['PATH'] = '/usr/local/bin:' + env.get('PATH', '')
-
-    full_command = f"{command} \"{input_path}\" "
-
-    logger.debug(f'Running mediainfo to generate MediaTrace XML: {full_command}')
-    output = subprocess.run(full_command, shell=True, capture_output=True)
-
-    return output
+    # Check if tool metadata checking is enabled
+    if output_path and command_config.command_dict['tools'][tool_name][f'check_{tool_name}'] == 'yes':
+        parse_function = parse_functions.get(tool_name)
+        if parse_function:
+            return parse_function(output_path)
+    
+    return None
 
 
-def write_to_csv(diff_dict, tool_name, writer):
-    for key, values in diff_dict.items():
-        actual_value, expected_value = values
-        writer.writerow({
-            'Metadata Tool': tool_name,
-            'Metadata Field': key,
-            'Expected Value': expected_value,
-            'Actual Value': actual_value
-        })
+def process_video_metadata(video_path, destination_directory, video_id, command_config):
+    """
+    Main function to process video metadata using multiple tools.
+    
+    Args:
+        video_path (str): Path to the input video file
+        destination_directory (str): Directory to store output files
+        video_id (str): Unique identifier for the video
+        command_config (object): Configuration object with tool settings
+        
+    Returns:
+        dict: Dictionary of metadata differences from various tools
+    """
+    # List of tools to process
+    tools = ['exiftool', 'mediainfo', 'mediatrace', 'ffprobe']
+    
+    # Store differences for each tool
+    metadata_differences = {}
+    
+    # Process each tool
+    for tool in tools:
+        # Run tool and get output path
+        output_path = tools_processing.run_tool_command(tool, video_path, destination_directory, video_id, command_config)
+        
+        # Check metadata and store differences
+        differences = check_tool_metadata(tool, output_path, command_config)
+        if differences:
+            metadata_differences[tool] = differences
+    
+    return metadata_differences
 
 
 def format_config_value(value, indent=0, is_nested=False):
@@ -243,7 +261,7 @@ The scripts will confirm that the digital files conform to predetermined specifi
         parser.error("the following arguments are required: paths")
 
     input_paths = [] if args.dryrun else args.paths
-    source_directories = validate_input_paths(input_paths, args.file)
+    source_directories = dir_setup.validate_input_paths(input_paths, args.file)
 
     # Determine save_config_type based on saveprofile
     if args.saveprofile == 'config':
@@ -285,90 +303,6 @@ def _generate_profile_filename(saveprofile):
                         f"{profile_type}_profile_{datetime.today().strftime('%Y-%m-%d_%H-%M-%S')}.yaml")
 
 
-def initialize_directory(source_directory):
-    """
-    Prepare the directory for processing by finding the video file 
-    and validating the filename.
-
-    Args:
-        source_directory (str): Path to the source directory
-
-    Returns:
-        tuple: (video_path, video_id, destination_directory) if successful
-        None if preparation fails
-    """
-    video_path = dir_setup.find_mkv(source_directory)
-
-    if video_path is None:
-        logger.warning(f"Skipping {source_directory} due to error.\n")
-        return None  # Indicates preparation failed
-
-    valid_filename = is_valid_filename(video_path)
-
-    if valid_filename is False:
-        logger.warning(f"Skipping {source_directory} due to error.\n")
-        return None  # Indicates preparation failed
-
-    logger.warning(f'Now processing {video_path}\n')
-
-    # outputs video_id (i.e. 'JPC_AV_05000')
-    video_id = os.path.splitext(os.path.basename(video_path))[0]
-
-    # Check to confirm directory is the same name as the video file name
-    dir_setup.check_directory(source_directory, video_id)
-
-    # Create 'destination directory' for qc outputs
-    destination_directory = dir_setup.make_qc_output_dir(source_directory, video_id)
-
-    # Moves vrecord files to subdirectory  
-    dir_setup.move_vrec_files(source_directory, video_id)
-
-    # Iterate through files in the directory to identify access file
-    access_file_found = None
-    for filename in os.listdir(source_directory):
-        if filename.lower().endswith('mp4'):
-            access_file_found = filename
-            logger.info("Existing access file found!\n")
-            break
-
-    return video_path, video_id, destination_directory, access_file_found
-
-
-def process_embedded_fixity(video_path):
-    """
-    Handles embedding stream fixity tags in the video file.
-    """
-    existing_tags = extract_tags(video_path)
-    if existing_tags:
-        existing_video_hash, existing_audio_hash = extract_hashes(existing_tags)
-    else:
-        existing_video_hash = None
-        existing_audio_hash = None
-
-    # Check if VIDEO_STREAM_HASH and AUDIO_STREAM_HASH MKV tags exist
-    if existing_video_hash is None or existing_audio_hash is None:
-        embed_fixity(video_path)
-    else:
-        logger.critical("Existing stream hashes found!")
-        if command_config.command_dict['outputs']['fixity']['overwrite_stream_fixity'] == 'yes':
-            logger.critical('New stream hashes will be generated and old hashes will be overwritten!')
-            embed_fixity(video_path)
-        elif command_config.command_dict['outputs']['fixity']['overwrite_stream_fixity'] == 'no':
-            logger.error('Not writing stream hashes to MKV\n')
-        elif command_config.command_dict['outputs']['fixity']['overwrite_stream_fixity'] == 'ask me':
-            # User input for handling existing stream hashes
-            while True:
-                user_input = input("Do you want to overwrite existing stream hashes? (yes/no): ")
-                if user_input.lower() in ["yes", "y"]:
-                    embed_fixity(video_path)
-                    break
-                elif user_input.lower() in ["no", "n"]:
-                    logger.debug('Not writing stream hashes to MKV')
-                    break
-                else:
-                    print("Invalid input. Please enter yes/no.")
-
-
 def process_fixity(source_directory, video_path, video_id):
     """
     Orchestrates the entire fixity process, including embedded and file-level operations.
@@ -396,182 +330,6 @@ def process_fixity(source_directory, video_path, video_id):
     # Verify stored checksum and write results
     if command_config.command_dict['outputs']['fixity']['check_fixity'] == 'yes':
         check_fixity(source_directory, video_id, actual_checksum=md5_checksum)
-
-
-def run_tool_command(tool_name, video_path, destination_directory, video_id, command_config):
-    """
-    Run a specific metadata extraction tool and generate its output file.
-    
-    Args:
-        tool_name (str): Name of the tool to run (e.g., 'exiftool', 'mediainfo')
-        video_path (str): Path to the input video file
-        destination_directory (str): Directory to store output files
-        video_id (str): Unique identifier for the video
-        command_config (object): Configuration object with tool settings
-        
-    Returns:
-        str or None: Path to the output file, or None if tool is not run
-    """
-    # Define tool-specific command configurations
-    tool_commands = {
-        'exiftool': {
-            'command': 'exiftool',
-            'config_key': 'run_exiftool'
-        },
-        'mediainfo': {
-            'command': 'mediainfo -f',
-            'config_key': 'run_mediainfo'
-        },
-        'mediatrace': {
-            'command': 'mediainfo --Details=1 --Output=XML',
-            'config_key': 'run_mediatrace'
-        },
-        'ffprobe': {
-            'command': 'ffprobe -v error -hide_banner -show_format -show_streams -print_format json',
-            'config_key': 'run_ffprobe'
-        }
-    }
-
-    # Check if the tool is configured to run
-    tool_config = tool_commands.get(tool_name)
-    if not tool_config:
-        logger.error(f"tool command is not configured correctly: {tool_name}")
-        return None
-
-    # Construct output path
-    output_path = os.path.join(destination_directory, f'{video_id}_{tool_name}_output.{_get_file_extension(tool_name)}')
-    
-    # Check if tool should be run based on configuration
-    if command_config.command_dict['tools'][tool_name][tool_config['config_key']] == 'yes':
-        if tool_name == 'mediatrace':
-            logger.debug(f"Creating {tool_name.capitalize()} XML file to check custom MKV Tag metadata fields:")
-        
-        # Run the tool command
-        run_command(tool_config['command'], video_path, '>', output_path)
-        
-        return output_path if os.path.isfile(output_path) else None
-    
-    return None
-
-def _get_file_extension(tool_name):
-    """
-    Get the appropriate file extension for each tool's output.
-    
-    Args:
-        tool_name (str): Name of the tool
-        
-    Returns:
-        str: File extension for the tool's output
-    """
-    extension_map = {
-        'exiftool': 'txt',
-        'mediainfo': 'txt',
-        'mediatrace': 'xml',
-        'ffprobe': 'txt'
-    }
-    return extension_map.get(tool_name, 'txt')
-
-def check_tool_metadata(tool_name, output_path, command_config):
-    """
-    Check metadata for a specific tool if configured.
-    
-    Args:
-        tool_name (str): Name of the tool
-        output_path (str): Path to the tool's output file
-        command_config (object): Configuration object with tool settings
-        
-    Returns:
-        dict or None: Differences found by parsing the tool's output, or None
-    """
-    # Mapping of tool names to their parsing functions
-    parse_functions = {
-        'exiftool': parse_exiftool,
-        'mediainfo': parse_mediainfo,
-        'mediatrace': parse_mediatrace,
-        'ffprobe': parse_ffprobe
-    }
-
-    # Check if tool metadata checking is enabled
-    if output_path and command_config.command_dict['tools'][tool_name][f'check_{tool_name}'] == 'yes':
-        parse_function = parse_functions.get(tool_name)
-        if parse_function:
-            return parse_function(output_path)
-    
-    return None
-
-def process_video_metadata(video_path, destination_directory, video_id, command_config):
-    """
-    Main function to process video metadata using multiple tools.
-    
-    Args:
-        video_path (str): Path to the input video file
-        destination_directory (str): Directory to store output files
-        video_id (str): Unique identifier for the video
-        command_config (object): Configuration object with tool settings
-        
-    Returns:
-        dict: Dictionary of metadata differences from various tools
-    """
-    # List of tools to process
-    tools = ['exiftool', 'mediainfo', 'mediatrace', 'ffprobe']
-    
-    # Store differences for each tool
-    metadata_differences = {}
-    
-    # Process each tool
-    for tool in tools:
-        # Run tool and get output path
-        output_path = run_tool_command(tool, video_path, destination_directory, video_id, command_config)
-        
-        # Check metadata and store differences
-        differences = check_tool_metadata(tool, output_path, command_config)
-        if differences:
-            metadata_differences[tool] = differences
-    
-    return metadata_differences
-
-
-def create_metadata_difference_report(metadata_differences, report_directory, video_id):
-    """
-    Create a CSV report of metadata differences.
-    
-    Args:
-        metadata_differences (dict): Dictionary of metadata differences from various tools
-        report_directory (str): Directory to save the report
-        video_id (str): Unique identifier for the video
-        
-    Returns:
-        str or None: Path to the created CSV report, or None if no differences
-    """
-    # If no metadata differences, log and return
-    if not metadata_differences:
-        logger.info("All specified metadata fields and values found, no CSV report written\n")
-        return None
-
-    # Prepare CSV file path
-    csv_name = f'{video_id}_metadata_difference.csv'
-    diff_csv_path = os.path.join(report_directory, csv_name)
-
-    try:
-        # Define CSV header and open file
-        fieldnames = ['Metadata Tool', 'Metadata Field', 'Expected Value', 'Actual Value']
-        
-        with open(diff_csv_path, 'w', newline='') as diffs_csv:
-            writer = csv.DictWriter(diffs_csv, fieldnames=fieldnames)
-            writer.writeheader()
-
-            # Write differences for each tool
-            tools_to_check = ['exiftool', 'mediainfo', 'mediatrace', 'ffprobe']
-            for tool in tools_to_check:
-                if tool in metadata_differences and metadata_differences[tool]:
-                    write_to_csv(metadata_differences[tool], tool, writer)
-
-        logger.info(f"Metadata difference report created: {diff_csv_path}\n")
-        return diff_csv_path
-
-    except IOError as e:
-        logger.critical(f"Error creating metadata difference CSV: {e}")
-        return None
 
 
 def process_qctools_output(video_path, source_directory, destination_directory, video_id, command_config, report_directory=None):
@@ -603,7 +361,7 @@ def process_qctools_output(video_path, source_directory, destination_directory, 
     
     try:
         # Run QCTools command
-        run_command('qcli -i', video_path, '-o', qctools_output_path)
+        tools_processing.run_command('qcli -i', video_path, '-o', qctools_output_path)
         logger.debug('')  # Add new line for cleaner terminal output
         results['qctools_output_path'] = qctools_output_path
 
@@ -626,72 +384,6 @@ def process_qctools_output(video_path, source_directory, destination_directory, 
         logger.critical(f"Error processing QCTools output: {e}")
 
     return results
-
-
-def process_access_file(video_path, source_directory, video_id, command_config):
-    """
-    Generate access file if configured and not already existing.
-    
-    Args:
-        video_path (str): Path to the input video file
-        source_directory (str): Source directory for the video
-        video_id (str): Unique identifier for the video
-        command_config (object): Configuration object with tool settings
-        
-    Returns:
-        str or None: Path to the created access file, or None
-    """
-    # Check if access file should be generated
-    if command_config.command_dict['outputs']['access_file'] != 'yes':
-        return None
-
-    access_output_path = os.path.join(source_directory, f'{video_id}_access.mp4')
-
-    try:
-        # Check if access file already exists
-        if os.path.isfile(access_output_path):
-            logger.critical(f"Access file already exists, not running ffmpeg\n")
-            return None
-
-        # Generate access file
-        make_access_file(video_path, access_output_path)
-        return access_output_path
-
-    except Exception as e:
-        logger.critical(f"Error creating access file: {e}")
-        return None
-
-
-def generate_final_report(video_id, source_directory, report_directory, destination_directory, command_config):
-    """
-    Generate final HTML report if configured.
-    
-    Args:
-        video_id (str): Unique identifier for the video
-        source_directory (str): Source directory for the video
-        report_directory (str): Directory containing report files
-        destination_directory (str): Destination directory for output files
-        command_config (object): Configuration object with tool settings
-        
-    Returns:
-        str or None: Path to the generated HTML report, or None
-    """
-    # Check if report should be generated
-    if command_config.command_dict['outputs']['report'] != 'yes':
-        return None
-
-    try:
-        html_report_path = os.path.join(source_directory, f'{video_id}_avspex_report.html')
-        
-        # Generate HTML report
-        write_html_report(video_id, report_directory, destination_directory, html_report_path)
-        
-        logger.info(f"HTML report generated: {html_report_path}\n")
-        return html_report_path
-
-    except Exception as e:
-        logger.critical(f"Error generating HTML report: {e}")
-        return None
 
 
 def process_video_outputs(video_path, source_directory, destination_directory, video_id, command_config, metadata_differences):
@@ -723,7 +415,7 @@ def process_video_outputs(video_path, source_directory, destination_directory, v
     if command_config.command_dict['outputs']['report'] == 'yes':
         report_directory = dir_setup.make_report_dir(source_directory, video_id)
         # Process metadata differences report
-        processing_results['metadata_diff_report'] = create_metadata_difference_report(
+        processing_results['metadata_diff_report'] = tools_processing.create_metadata_difference_report(
                 metadata_differences, report_directory, video_id
             )
     else:
@@ -837,7 +529,7 @@ def process_single_directory(source_directory):
     with create_processing_timer() as timer:
         try:
             # Call the new prep_directory function
-            init_dir_result = initialize_directory(source_directory)
+            init_dir_result = dir_setup.initialize_directory(source_directory)
             if init_dir_result is None:
                 return  # Skip to the next source_directory if preparation failed
