+++ src/AV_Spex/utils/dir_setup.py
@@ -138,3 +138,101 @@ def make_report_dir(source_directory, video_id):
     logger.debug(f'Report files will be written to {report_directory}\n')
 
     return report_directory
+
+
+def is_valid_filename(video_filename):
+    '''
+    Locates approved values for the file name, stored in key:value pairs under 'filename_values' in config/config.yaml.
+    The file name pattern varies and is constructed dynamically depending on the provided config.
+    '''
+    valid_filename = False
+    
+    # Reads filename_values from config.yaml into dictionary approved_values
+    approved_values = config_path.config_dict['filename_values']
+
+    # Get only the base filename (not the full path)
+    base_filename = os.path.basename(video_filename)
+
+    # Build the dynamic regex pattern based on the keys in the config
+    pattern_parts = [
+        re.escape(approved_values['Collection']),
+        re.escape(approved_values['MediaType']),
+        approved_values['ObjectID']  # ObjectID can contain a regex, so no escaping
+    ]
+    
+    # Check if 'DigitalGeneration' is part of the convention and include it in the pattern if present
+    if 'DigitalGeneration' in approved_values:
+        pattern_parts.append(re.escape(approved_values['DigitalGeneration']))
+    
+    # Append the file extension
+    file_extension = re.escape(approved_values['FileExtension'])
+    
+    # Construct the complete pattern, joining the parts and accounting for the file extension
+    pattern = r'^{0}\.{1}$'.format('_'.join(pattern_parts), file_extension)
+    
+    # Check if the filename matches the pattern
+    if re.match(pattern, base_filename, re.IGNORECASE):
+        logger.debug(f"The file name '{base_filename}' is valid.\n")
+        valid_filename = True
+    else:
+        logger.critical(f"The file name '{base_filename}' is not valid.\n")
+        valid_filename = False
+
+    return valid_filename
+
+
+if sys.argv[0] == __file__:
+    if len(sys.argv) != 2:
+        print("Usage: python filename_check.py /path/to/your/directory")
+    else:
+        video_path = sys.argv[1]
+        valid_filename = is_valid_filename(video_path)
+
+
+def initialize_directory(source_directory):
+    """
+    Prepare the directory for processing by finding the video file 
+    and validating the filename.
+
+    Args:
+        source_directory (str): Path to the source directory
+
+    Returns:
+        tuple: (video_path, video_id, destination_directory) if successful
+        None if preparation fails
+    """
+    video_path = find_mkv(source_directory)
+
+    if video_path is None:
+        logger.warning(f"Skipping {source_directory} due to error.\n")
+        return None  # Indicates preparation failed
+
+    valid_filename = is_valid_filename(video_path)
+
+    if valid_filename is False:
+        logger.warning(f"Skipping {source_directory} due to error.\n")
+        return None  # Indicates preparation failed
+
+    logger.warning(f'Now processing {video_path}\n')
+
+    # outputs video_id (i.e. 'JPC_AV_05000')
+    video_id = os.path.splitext(os.path.basename(video_path))[0]
+
+    # Check to confirm directory is the same name as the video file name
+    check_directory(source_directory, video_id)
+
+    # Create 'destination directory' for qc outputs
+    destination_directory = make_qc_output_dir(source_directory, video_id)
+
+    # Moves vrecord files to subdirectory  
+    move_vrec_files(source_directory, video_id)
+
+    # Iterate through files in the directory to identify access file
+    access_file_found = None
+    for filename in os.listdir(source_directory):
+        if filename.lower().endswith('mp4'):
+            access_file_found = filename
+            logger.info("Existing access file found!\n")
+            break
+
+    return video_path, video_id, destination_directory, access_file_found
